no data view

+++ views/contact_us/index.php
@@ -7,29 +7,35 @@ use yii\helpers\Html; ?>
   </div>
   <div class="col-lg">
     <?= $this->render('../shared/content_header', ['showNewButton'=> false]);?>
-    <table id="t01">
-      <tr>
-        <th>Nome</th>
-        <th>Descrição</th>
-        <th>Email</th>
-        <th>Número</th>
-        <th></th>
-      </tr>
-      <?php foreach ($contact_us as $contact_us_data): ?>
-        <tr>
-          <td><?= $contact_us_data->name ?></td>
-            <td> <?= yii\helpers\BaseStringHelper::truncate($contact_us_data->description, 40) ?></td>
-          <td><?= $contact_us_data->email ?></td>
-          <td><?= $contact_us_data->number ?></td>
-          <td class="td_button">
-            <a href="/index.php?r=contact_us/show&contact_us_id=<?= $contact_us_data->id ?>"><i class="bi bi-eye-fill"></i></a>
-              <a href="/index.php?r=contact_us/delete&contact_us_id=<?= $contact_us_data->id ?>" data-confirm="Are you sure?">
-                  <i class="bi bi-trash-fill"></i>
-              </a>
-          </td>
-        </tr>
-      <?php endforeach;?>
-    </table>
+
+    <?php if(count($contact_us) > 0): ?>
+        <table id="t01">
+            <tr>
+                <th>Nome</th>
+                <th>Descrição</th>
+                <th>Email</th>
+                <th>Número</th>
+                <th></th>
+            </tr>
+          <?php foreach ($contact_us as $contact_us_data): ?>
+              <tr>
+                  <td><?= $contact_us_data->name ?></td>
+                  <td> <?= yii\helpers\BaseStringHelper::truncate($contact_us_data->description, 40) ?></td>
+                  <td><?= $contact_us_data->email ?></td>
+                  <td><?= $contact_us_data->number ?></td>
+                  <td class="td_button">
+                      <a href="/index.php?r=contact_us/show&contact_us_id=<?= $contact_us_data->id ?>"><i class="bi bi-eye-fill"></i></a>
+                      <a href="/index.php?r=contact_us/delete&contact_us_id=<?= $contact_us_data->id ?>" data-confirm="Are you sure?">
+                          <i class="bi bi-trash-fill"></i>
+                      </a>
+                  </td>
+              </tr>
+          <?php endforeach;?>
+        </table>
+    <?php else: ?>
+        <div class="empety_repository"> <i class="bi bi-info-circle-fill"></i> Atualmente não existem ficheiros nesse local.</div>
+    <?php endif; ?>
+
   </div>
 
 </div>
